Refactor index page and iCal builder to own files
